# src/test/resources/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Take home tech task
  version: v1.0.0
  description: |-
    We want you to create a REST API for Eagle Bank which conforms to this OpenAPI specification which allows a user to create, fetch, update and delete a bank account and deposit or withdraw money from the account. These will be stored as transactions against a bank account which be retrieved but not modified or deleted.
...
paths:
...
  /v1/accounts/{accountNumber}/transactions:
    post:
      tags:
        - transaction
      description: Create a transaction
      operationId: createTransaction
      parameters:
        - name: accountNumber
          in: path
          description: Account number of the bank account
          required: true
          schema:
            type: string
            pattern: ^01\d{6}$
      requestBody:
        description: Create a new transaction
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransactionRequest'
        required: true
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Transaction has been created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        '400':
          description: Invalid details supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestErrorResponse'
        '401':
          description: Access token is missing or invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '403':
          description: The user is not allowed to delete the bank account details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: Bank account was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '422':
          description: Insufficient funds to process transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      tags:
        - transaction
      description: List transactions
      operationId: listAccountTransaction
      parameters:
        - name: accountNumber
          in: path
          description: Account number of the bank account
          required: true
          schema:
            type: string
            pattern: ^01\d{6}$
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The list of transaction details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTransactionsResponse'
        '400':
          description: The request didn't supply all the necessary data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BadRequestErrorResponse"
        '401':
          description: Access token is missing or invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '403':
          description: The user is not allowed to access the transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '404':
          description: Bank account was not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        '500':
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    CreateTransactionRequest:
      type: object
      required:
        - amount
        - currency
        - type
      properties:
        amount:
          type: number
          format: double
          minimum: 0.00
          maximum: 10000.00
          description: "Currency amount with up to two decimal places"
          examples:
            - 10.99
            - 1000.00
        currency:
          type: string
          enum:
            - "GBP"
        type:
          type: string
          enum: 
            - "deposit"
            - "withdrawal"
        reference:
          type: string
    ListTransactionsResponse:
      type: object
      required:
        - transactions
      properties:
        transactions:
          type: array
          items:
             $ref: "#/components/schemas/TransactionResponse"
    TransactionResponse:
      type: object
      required:
        - id
        - amount
        - currency
        - type
        - createdTimestamp
      properties:
        id: 
          type: string
          pattern: ^tan-[A-Za-z0-9]$
          examples:
            - tan-123abc
        amount:
          type: number
          format: double
          minimum: 0.00
          maximum: 10000.00
        currency:
          type: string
          enum:
            - "GBP"
        type:
          type: string
          enum: 
            - "deposit"
            - "withdrawal"
        reference:
          type: string
        userId:
          type: string
          format: ^usr-[A-Za-z0-9]+$
          examples: 
            - usr-abc123
        createdTimestamp:
          type: string
          format: 'date-time'
...
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    BadRequestErrorResponse:
      type: object
      required:
        - message
        - details
      properties:
        message:
          type: string
        details:
          type: array
          items:
            type: object
            required:
              - field
              - message
              - type
            properties:
              field:
                type: string
              message:
                type: string
              type:
                type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT